Asset decomposition
